move tests to integration test
